requirements.txt
pandas
requests
scikit-learn
boto
